requirements.txt
google-genai==2.11.0
feedparser==6.0.11
